server/requirements.txt
Flask==1.0.2
mysql-connector
requests
